test: add minimal repro for dual-source cad fallback
